Lesson 6:03/08/20 13:30 1) Доделать собаку;solved++ 2) Разобраться с nextDouble и nextLong для Random;+- float тоже не работает - остается только int? 3) Разобраться с последней задачей; 4) Что придумать вместо геттеров, чтобы обозначать для животных предельные расстояния? 5) Сделать рандомный выбор между run, jump и swim?

+++ DZ_1/src/Animal.java
@@ -1,11 +1,11 @@
 public abstract class Animal {
     protected int obstacleLength;
-    protected long obstacleHeight;
+    protected float obstacleHeight;
     protected String name;
 
     public Animal () {}
 
-    public Animal (int obstacleLength, long obstacleHeight, String name) {
+    public Animal (int obstacleLength, float obstacleHeight, String name) {
         this.obstacleLength = obstacleLength;
         this.obstacleHeight = obstacleHeight;
         this.name = name;
@@ -16,7 +16,7 @@ public abstract class Animal {
         return obstacleLength;
     }
 
-    public long getObstacleHeight() {
+    public float getObstacleHeight() {
         return obstacleHeight;
     }
 
@@ -35,7 +35,7 @@ public abstract class Animal {
 
     public abstract void swim (int obstacleLength);
 
-    public abstract void jump (long obstacleHeight);
+    public abstract void jump (float obstacleHeight);
 
     public abstract void run (int obstacleLength);
